Another adjustment to APCA contrast
Now use the absolute value of the unclamped channel because values can 
be negative. Don't want rogue NaNs out there because negative numbers 
cannot have fractional exponents. The short of it (rather than 
explaining in detail the math) is that in the process of multiplying the 
absolute value it would end up being positive anyway. This is -in 
theory- after a long day so might change tomorrow. :)